Created Landing/Home page, Account creation
Created the landing page for the application as well as the home page. Account creation is now possible with the ability to register accounts with security checks to ensure that users are creating accounts that meet a certain criteria (minimum password length, username requirements). Once users log in, a home page was created to redirect users to. Users can also log out.